Split event publication into nested sub-topics for MQTT
This changes from posting the full JSON payload to the topic to
enumerated sub-topics nested under the vehicle ID. This makes it
possible to subscribe to a specific item for updates, while allowing
for multiple instances of the simulator to publish events through
the same broker (e.g. in fleet simulation cases).