Add driver_changed webhook
Add basic support for journeys being taken over by different drivers.
This can occur either in cases where a journey is paused and a driver
physically changes (where the driver is a natural person), or in the
case of an autonomous driving system, in which the vehicle takes
over driving functions at certain automation levels.